requirements.txt
activations==0.1.0
dassl.egg==info
django_cache_utils==0.7.2
flash_attn==2.6.3
ftfy==6.2.3
huggingface_hub==0.24.6
ImageBind==0.0.1.dev0
llava==1.7.0.dev0
numpy==2.1.1
openai==1.48.0
pandas==2.2.3
Pillow==10.4.0
pytorch_utils==0.5.5
regex==2024.7.24
Requests==2.32.3
scipy==1.14.1
sentence_transformers==3.0.1
setuptools==72.1.0
timm==1.0.9
torch==2.2.2
torch==2.1.2
torchvision==0.16.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.40.0.dev0
yacs==0.1.8
